+++ test/nft-marketplace/nft-marketplace.test.ts
@@ -1,10 +1,10 @@
 import { web3, subContractId, binToHex, encodeU256, addressFromContractId, sleep, ONE_ALPH } from '@alephium/web3'
 import { PrivateKeyWallet } from '@alephium/web3-wallet'
 import { getSigners } from '@alephium/web3-test'
-import { NFTCollectionHelper } from '../shared/nft-collection'
-import { NFTMarketplace } from '../shared/nft-marketplace'
-import { NFTListingInstance, NFTMarketPlace, NFTMarketPlaceInstance } from '../artifacts/ts'
-import { contractExists } from '../shared'
+import { NFTCollectionHelper } from '../../shared/nft-collection'
+import { NFTMarketplace } from '../../shared/nft-marketplace'
+import { NFTListingInstance, NFTMarketPlaceInstance } from '../../artifacts/ts'
+import { contractExists } from '../../shared'
 import { Balance } from '@alephium/web3/dist/src/api/api-alephium'
 
 describe('nft marketplace', function() {
